requirements.txt
streamlit>=1.33.0
langgraph>=0.1.8
langchain>=0.1.0
langchain-core>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=5.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0 